# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budget Time-Machine production planning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signupProducer
      tags: [auth]
      summary: Create a producer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProducerSignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/signin:
    post:
      operationId: signinProducer
      tags: [auth]
      summary: Sign in a producer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProducerCredentialsInput"
      responses:
        "200":
          description: Signed-in producer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ProducerUser:
      type: object
      properties:
        userId: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
      required: [userId, email, displayName]
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/ProducerUser"
      required: [user]
    ProducerCredentialsInput:
      type: object
      properties:
        email: { type: string, format: email, maxLength: 255 }
        password: { type: string, minLength: 8, maxLength: 200 }
      required: [email, password]
    ProducerSignupInput:
      allOf:
        - $ref: "#/components/schemas/ProducerCredentialsInput"
        - type: object
          properties:
            displayName: { type: string, minLength: 1, maxLength: 80 }
          required: [displayName]
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...