# lib/api-zod/src/generated/types/producerUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Budget Time-Machine production planning API
 * OpenAPI spec version: 0.1.0
 */

export interface ProducerUser {
  userId: string;
  email: string;
  displayName: string;
}